Move test functionality out of product code
